Quiesce clang 3.0's warning regarding PRId32 and SInt32, which is typedef'd as 'long' on non-LP64 Mac OS X.
git-svn-id: https://plcrashreporter.googlecode.com/svn/trunk@391 25172300-ee46-11dd-abe2-393a09110dd0

+++ Source/PLCrashLogWriter.m
@@ -370,7 +370,7 @@ plcrash_error_t plcrash_log_writer_init (plcrash_log_writer_t *writer, NSString 
         }
 
         /* Compose the string */
-        asprintf(&writer->system_info.version, "%" PRId32 ".%" PRId32 ".%" PRId32, major, minor, bugfix);
+        asprintf(&writer->system_info.version, "%" PRId32 ".%" PRId32 ".%" PRId32, (int32_t)major, (int32_t)minor, (int32_t)bugfix);
     }
 #else
 #error Unsupported Platform
